requirements.txt
Flask==3.1.0
numpy==2.0.2
requests==2.32.3
tensorflow==2.18.0
pillow==11.0.0
firebase-admin==6.6.0